refactor(@covalent/experimental): Replace ng-packagr with cli command
