+++ scripts/build-release
@@ -38,7 +38,7 @@ cp -r deploy/platform/http src/platform/node_modules/@covalent
 # BUILD: @covalent/flavored-markdown primary entrypoint
 ./node_modules/.bin/ng-packagr -p src/platform/flavored-markdown/ng-package.js
 # BUILD: @covalent/experimental primary entrypoint
-./node_modules/.bin/ng-packagr -p src/platform/experimental/ng-package.js
+./node_modules/@angular/cli/bin/ng build covalent-experimental
 
 # HACK (Remove in future): remove remnants of platform/node_modules was used for building purposes
 rm -rf src/platform/node_modules
