+++ src/platform/experimental/karma.conf.js
@@ -1,0 +1,4 @@
+// Karma configuration file, see link for more information
+// https://karma-runner.github.io/1.0/config/configuration-file.html
+const defaultConfig = require('../../karma-config-default');
+module.exports = defaultConfig;
