+++ src/platform/experimental/tsconfig.lib.json
@@ -1,0 +1,6 @@
+{
+  "extends": "../../tsconfig.app.json",
+  "compilerOptions": {
+    "baseUrl": "."
+  }
+}
