+++ src/platform/experimental/tsconfig.spec.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.spec.json",
+  "files": [
+    "test.ts"
+  ],
+  "include": [
+    "**/*.spec.ts",
+    "**/*.d.ts"
+  ]
+}
